src/ai-voice/requirements.txt
# ==========================================
# Chipi TTS 프로젝트 의존성
# Python 3.7.3 호환 버전 (라즈베리파이 제로 WH)
# ==========================================

# Core Dependencies
python-dotenv>=0.19.0,<1.0.0  # 환경 변수 관리 (Python 3.7 호환)
requests>=2.25.0,<3.0.0       # HTTP 요청 (Python 3.7 호환)

# LLM & AI
# openai 0.28.x는 Python 3.7.3에서 동작, 1.x는 Python 3.8+ 필요
# 라즈베리파이 제로 WH에서는 0.28.x 사용 권장
openai>=0.28.0,<1.0.0         # Azure OpenAI API (Python 3.7 호환)

# Speech & Audio
# azure-cognitiveservices-speech는 Python 3.7에서도 동작하지만 버전 확인 필요
azure-cognitiveservices-speech>=1.20.0,<2.0.0  # Azure Speech SDK (STT/TTS)
pygame>=2.0.0,<3.0.0          # 음성 재생 (Python 3.7 호환)

# Database
psycopg2-binary>=2.8.0,<3.0.0  # PostgreSQL 드라이버 (Python 3.7 호환)

# Development (선택사항)
# pytest>=7.0.0                # 테스트 프레임워크
# black>=23.0.0                # 코드 포매팅
# pylint>=2.0.0                # 코드 린팅
